Remove ez-conf-lib:exe variable
This package used to provide a variable `exe` containing the absolute
path to the `ez-conf-lib` script inside the current opam switch. The
problem with using an absolute path here is it means the switch is not
portable. This causes problems for dune packag emanagement where all
packages are first installed to a temporary sandbox and then moved to a
final location afterwards.

Clients of ez-conf-lib are updated to refer to the script without using
the exe variable in: https://github.com/ocaml/opam-repository/pull/27759

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>